feat: support spoilers
Per https://spec.matrix.org/latest/client-server-api/#mroommessage-msgtypes the element span is to be treated almost identical to font, with the addition of data-mx-spoiler. So treating them identical seems like a good way to save space to me.

In addition, in IRC having foreground and background the same color is an agreed upon way to treat something as a spoiler, see https://modern.ircdocs.horse/formatting.html#spoilers

I dont usually code in python and this is untested. Please dont shoot me if this doesn't work.